+++ backend/database/migrations/rename_is_paid_to_is_premium.sql
@@ -1,0 +1,1 @@
+ALTER TABLE channels CHANGE is_paid is_premium TINYINT(1) DEFAULT 0;